Move error types to errors.go
